docs: Winning.js의 error 검증 조건 error.js 파일로 변경
