+++ __tests__/Lotto.test.js
@@ -1,6 +1,5 @@
 import Lotto from '../src/tests/Lotto';
-import { validateWinningNumbers } from '../src/tests/Winning';
-import { ErrorMessages } from '../src/tests/error';
+import { ErrorMessages, validateWinningNumbers } from '../src/tests/error';
 
 describe('로또 기능 테스트', () => {
 	test('로또 1장의 가격은 1000원이다', () => {
